src/tools/integration_tools.py
"""
Integration Tools for AI Business Orchestra
External service integrations for enhanced functionality
"""

import os
from typing import Dict, Any, List
from crewai_tools import BaseTool

class EmailTool(BaseTool):
    """Tool for email operations (Gmail integration)"""
    
    name: str = "Email Tool"
    description: str = "Send emails, manage email communications and notifications"
    
    def _run(self, recipient: str, subject: str, body: str) -> str:
        """Send email (simulation mode)"""
        if os.getenv("USE_SIMULATION", "True").lower() == "true":
            return f"📧 Email sent to {recipient}\nSubject: {subject}\nBody: {body[:100]}..."
        
        # TODO: Implement real Gmail API integration
        return "Email tool requires Gmail API configuration"

class CalendarTool(BaseTool):
    """Tool for calendar operations (Google Calendar/Outlook)"""
    
    name: str = "Calendar Tool"
    description: str = "Schedule meetings, manage calendar events and availability"
    
    def _run(self, action: str, details: Dict[str, Any]) -> str:
        """Manage calendar events (simulation mode)"""
        if os.getenv("USE_SIMULATION", "True").lower() == "true":
            if action == "schedule":
                return f"📅 Meeting scheduled: {details.get('title', 'Meeting')} on {details.get('date', 'TBD')}"
            elif action == "check_availability":
                return f"✅ Available slots found for {details.get('date', 'requested date')}"
        
        # TODO: Implement real Calendar API integration
        return "Calendar tool requires API configuration"

class DocumentTool(BaseTool):
    """Tool for document management (Google Drive/SharePoint)"""
    
    name: str = "Document Tool"
    description: str = "Create, edit, and manage documents and files"
    
    def _run(self, action: str, document_info: Dict[str, Any]) -> str:
        """Manage documents (simulation mode)"""
        if os.getenv("USE_SIMULATION", "True").lower() == "true":
            if action == "create":
                return f"📄 Document created: {document_info.get('name', 'New Document')}"
            elif action == "share":
                return f"🔗 Document shared with {document_info.get('recipients', 'team')}"
        
        # TODO: Implement real Drive/SharePoint API integration
        return "Document tool requires API configuration"

class VideoConferenceTool(BaseTool):
    """Tool for video conferencing (Zoom/Teams integration)"""
    
    name: str = "Video Conference Tool"
    description: str = "Schedule and manage video conferences and online meetings"
    
    def _run(self, action: str, meeting_details: Dict[str, Any]) -> str:
        """Manage video conferences (simulation mode)"""
        if os.getenv("USE_SIMULATION", "True").lower() == "true":
            if action == "create_meeting":
                return f"🎥 Video meeting created: {meeting_details.get('title', 'Team Meeting')}"
            elif action == "send_invite":
                return f"📨 Meeting invite sent to {len(meeting_details.get('attendees', []))} attendees"
        
        # TODO: Implement real Zoom/Teams API integration
        return "Video conference tool requires API configuration"

class HRSystemTool(BaseTool):
    """Tool for HR system integration (ATS/HRIS)"""
    
    name: str = "HR System Tool"
    description: str = "Integrate with HR systems for candidate tracking and employee management"
    
    def _run(self, action: str, hr_data: Dict[str, Any]) -> str:
        """Manage HR system operations (simulation mode)"""
        if os.getenv("USE_SIMULATION", "True").lower() == "true":
            if action == "post_job":
                return f"💼 Job posted: {hr_data.get('title', 'Position')} on HR platforms"
            elif action == "track_candidate":
                return f"👤 Candidate {hr_data.get('name', 'Unknown')} added to tracking system"
        
        # TODO: Implement real ATS/HRIS integration
        return "HR system tool requires API configuration"

# Tool registry for easy access
AVAILABLE_TOOLS = {
    'email': EmailTool(),
    'calendar': CalendarTool(),
    'document': DocumentTool(),
    'video_conference': VideoConferenceTool(),
    'hr_system': HRSystemTool()
}

def get_tool(tool_name: str) -> BaseTool:
    """Get a specific tool by name"""
    return AVAILABLE_TOOLS.get(tool_name)

def get_all_tools() -> List[BaseTool]:
    """Get all available tools"""
    return list(AVAILABLE_TOOLS.values())


# Tool aliases for Business crew compatibility
GmailIntegrationTool = EmailTool
ZoomAPITool = VideoConferenceTool
GoogleDriveIntegrationTool = DocumentTool
CalendarIntegrationTool = CalendarTool
GoogleSheetsIntegrationTool = DocumentTool
