pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-business-orchestra"
version = "0.1.0"
description = "AI Business Orchestra - Multi-Agent Business Automation System"
authors = [
    { name = "UMUR KIZILDAS", email = "umur@kizildas.com" }
]
dependencies = [
    "crewai>=0.80.0",
    "crewai[tools]>=0.80.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "smtplib-ssl>=1.0.0",
    "email-validator>=2.0.0"
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[project.scripts]
business-orchestra = "business_orchestra_main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
